+++ pakyow-routing/spec/features/routing_spec.rb
@@ -73,14 +73,6 @@ RSpec.describe "routing requests" do
       call("/", method: :head)
       expect(Pakyow.apps.first.called).to be(true)
     end
-
-    it "responds with an empty body" do
-      expect(call("/", method: :head)[2].length).to eq(0)
-    end
-
-    it "responds with a valid content length header" do
-      expect(call("/", method: :head)[1]).to include("content-length" => "5")
-    end
   end
 
   context "when a default route is specified" do